fix byte overflow

+++ core/util/packed/bulkOperation12.go
@@ -43,11 +43,11 @@ func (op *BulkOperationPacked12) decodeByteToInt(blocks []byte, values []int32, 
 		blocksOffset++
 		byte1 := blocks[blocksOffset]
 		blocksOffset++
-		values[valuesOffset] = int32((byte0 << 4) | byte(uint8(byte1) >> 4))
+		values[valuesOffset] = int32((int64(byte0) << 4) | int64(uint8(byte1) >> 4))
 		valuesOffset++
 		byte2 := blocks[blocksOffset]
 		blocksOffset++
-		values[valuesOffset] = int32(((byte1 & 15) << 8) | byte2)
+		values[valuesOffset] = int32((int64(byte1 & 15) << 8) | int64(byte2))
 		valuesOffset++
 	}
 }
@@ -83,11 +83,11 @@ func (op *BulkOperationPacked12) decodeByteToLong(blocks []byte, values []int64,
 		blocksOffset++
 		byte1 := blocks[blocksOffset]
 		blocksOffset++
-		values[valuesOffset] = int64((byte0 << 4) | byte(uint8(byte1) >> 4))
+		values[valuesOffset] = int64((int64(byte0) << 4) | int64(uint8(byte1) >> 4))
 		valuesOffset++
 		byte2 := blocks[blocksOffset]
 		blocksOffset++
-		values[valuesOffset] = int64(((byte1 & 15) << 8) | byte2)
+		values[valuesOffset] = int64((int64(byte1 & 15) << 8) | int64(byte2))
 		valuesOffset++
 	}
 }
